Move eosio => ship type conversion to type_convert.hpp
